[evol] HttpReqError return error message as a HTML string like: <head> <title>Error response</title> </head> <body> <h1>Error response</h1> <p>Error code 400. <p>Message: <pre> missing 'ca' option</pre> . <p>Error code explanation: 400 = Bad request syntax or unsupported method. </body>
which is not ideal to parse
So I add a method to return such error message as a json string like:
{"message": "missing 'ca' option", "code": 400}

The modification is done in a way to keep full compatibility with previous code.
By default, error message is always formatted as HTML code.
You need to set *json=True* parameter when instanciating HttpReqError exception to return json string



git-svn-id: https://rcs.lan.proformatique.com/svn/xivo/trunk@10767 dc85baef-7018-0410-bfbd-8a639da648bb