add token renewer
There's more and more XiVO services that needs the "token renewal"
feature, so put the code in common here.

The implementation is similar to what was previously found in
xivo-agid/xivo-dird-phoned, i.e. it uses daemon thread, and stop()
doesn't join the renewer thread, with the small race-condition fixed.
That said, the implementation could be changed without changing the API.