consul_helper: make service discovery auto even more auto
if the service_discovery_address is auto the service_discovery_interface
will be used to find the ip address. If the specified interface does not
have an ip address, all interfaces starting with "eth" will be checked
for an ip address. The first address that is found will be returned. If
all "eth" interface fail, "lo" is used. If "lo" has no ip address???
127.0.0.1 is returned.

The problem is that some services fail to start when not using eth0
since our default values are auto/eth0. This will avoid raising
exceptions when the service is starting and publish 127.0.0.1 for
service discovery if all else fails.

ticket #6160