ServiceFinder: remove the remote_tokens args
consul can be configured to allow service discovery with the anonymous
tokens using the following command.

consul-cli  --token-file=/var/lib/consul/master_token --ssl
--ssl-verify=false acl update anonymous --rule='service:xivo-:read'

This allows the configuration of each service to be much simpler by not
having to configure each datacerter and it's service discovery token for
each service on each node.