auth: add a @no_auth verifier for URL that do not require authentication
this is usefull for resources that have multiple routes and one or more of them do not require a token.

A resource that is used for forgetten passwords for example will have a call that will trigger the
mail to be sent to the user. This call cannot require a token since the email is forgotten.

The call that will set the new password on the same URL will have a token that has been sent
to the user in the forgotten password email