tenant helpers: raise UnauthorizedTenant if the user does not exist
this case happens when using the xivo_admin or xivo_service auth backends.
If a tenant was specified using the Wazo-Tenant header and the token does not
include that tenant, it will raise an InvalidUser exception while trying to find
this tokens user_uuid.

If we get to that point, it means that the "user" cannot use the specified token