improve tenant cache to avoid duplicate requests
reason: tenant check_against_token can be implemented by trying to list
this specific tenant. If user has access to see it, then it's because
it's part of his sub tenant. Moreover, by doing this way, it will
initialized cache with a value that will prevent a second call to
visible_tenants(tenant_uuid=<tenant-uuid-to-verify)