token_renewer: allow to emit stop without calling stop
reason: used by services that create token only to fetch tenant_uuid and
automatically stop token_renewer with the callback. In the case to avoid
to block process if service is killed, we always try to emit stop

Moreover, callback can be executed by the main thread (the first
renew_token) or by a separate thread. That's why stop() cannot be called
directly